+++ test/packages_handler_test.dart
@@ -2,12 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'dart:convert';
 import 'dart:io';
-import 'dart:typed_data';
 
-import 'package:package_config/package_config.dart';
-import 'package:path/path.dart' as p;
 import 'package:shelf/shelf.dart';
 import 'package:shelf_packages_handler/shelf_packages_handler.dart';
 import 'package:test/test.dart';
@@ -40,29 +36,16 @@ void main() {
           await response.readAsString(), contains('Handler packagesHandler'));
     });
 
-    group('with a package config', () {
-      PackageConfig packageConfig;
+    group('with a package map', () {
+      Handler handler;
 
-      setUp(() async {
-        packageConfig = await loadPackageConfigUri(
-            Uri.file(p.absolute('.dart_tool/package_config.json')),
-            loader: (_) async => Uint8List.fromList(utf8.encode('''
-{
-  "configVersion": 2,
-  "packages": [
-    {
-      "name": "foo",
-      "rootUri": "${p.toUri('$dir/foo')}",
-      "packageUri": "",
-      "languageVersion": "2.2"
-    }
-  ]
-}
-''')));
+      setUp(() {
+        handler = packagesHandler(packageMap: {
+          'foo': Uri.file('$dir/foo/'),
+        });
       });
 
       test('looks up a real file', () async {
-        var handler = packagesHandler(packageConfig: packageConfig);
         var request =
             Request('GET', Uri.parse('http://example.com/foo/foo.dart'));
         var response = await handler(request);
@@ -71,7 +54,6 @@ void main() {
       });
 
       test('404s for a nonexistent package', () async {
-        var handler = packagesHandler(packageConfig: packageConfig);
         var request =
             Request('GET', Uri.parse('http://example.com/bar/foo.dart'));
         var response = await handler(request);
@@ -81,7 +63,6 @@ void main() {
       });
 
       test('404s for a nonexistent file', () async {
-        var handler = packagesHandler(packageConfig: packageConfig);
         var request =
             Request('GET', Uri.parse('http://example.com/foo/bar.dart'));
         var response = await handler(request);